add all neccessary classes so far
