+++ backend/src/test/java/com/engineering/thesis/backend/BackendApplicationTests.java
@@ -2,19 +2,12 @@ package com.engineering.thesis.backend;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.jdbc.Sql;
-import org.springframework.test.context.jdbc.SqlGroup;
-import org.springframework.test.context.jdbc.Sql.ExecutionPhase;
-
-@SqlGroup({
-		@Sql(executionPhase = ExecutionPhase.AFTER_TEST_METHOD ,scripts = "classpath:delete_data_after_test.sql")
-})
 
 @SpringBootTest
 class BackendApplicationTests {
 
 	@Test
-	public void initialTest(){
+	public void ContextLoads(){
 	}
 }
