--- backend/src/test/java/com/engineering/thesis/backend/ControllerUnitTests/Appointment/Post.java
@@ -1,4 +1,0 @@
-package com.engineering.thesis.backend.ControllerUnitTests.Appointment;
-
-public class Post {
-}
