--- backend/src/test/java/com/engineering/thesis/backend/ControllerUnitTests/Authentification/Login.java
@@ -1,4 +1,0 @@
-package com.engineering.thesis.backend.ControllerUnitTests.Authentification;
-
-public class Login {
-}
