--- backend/src/test/java/com/engineering/thesis/backend/ControllerUnitTests/Patient/GetAll.java
@@ -1,4 +1,0 @@
-package com.engineering.thesis.backend.ControllerUnitTests.Patient;
-
-public class GetAll {
-}
