--- backend/src/test/java/com/engineering/thesis/backend/ControllerUnitTests/Patient/GetById.java
@@ -1,4 +1,0 @@
-package com.engineering.thesis.backend.ControllerUnitTests.Patient;
-
-public class GetById {
-}
