--- backend/src/test/java/com/engineering/thesis/backend/ControllerUnitTests/PatientMedicalData/GetById.java
@@ -1,4 +1,0 @@
-package com.engineering.thesis.backend.ControllerUnitTests.PatientMedicalData;
-
-public class GetById {
-}
