--- backend/src/test/java/com/engineering/thesis/backend/ControllerUnitTests/User/Get.java
@@ -1,4 +1,0 @@
-package com.engineering.thesis.backend.ControllerUnitTests.User;
-
-public class Get {
-}
