--- backend/src/test/java/com/engineering/thesis/backend/ControllerUnitTests/User/Update.java
@@ -1,4 +1,0 @@
-package com.engineering.thesis.backend.ControllerUnitTests.User;
-
-public class Update {
-}
