+++ backend/src/test/java/com/engineering/thesis/backend/ServiceUnitTests/Appointment/GetById.java
@@ -1,0 +1,5 @@
+package com.engineering.thesis.backend.ServiceUnitTests.Appointment;
+
+public class GetById {
+
+}
