+++ backend/src/test/java/com/engineering/thesis/backend/ServiceUnitTests/Appointment/Post.java
@@ -1,0 +1,4 @@
+package com.engineering.thesis.backend.ServiceUnitTests.Appointment;
+
+public class Post {
+}
