+++ backend/src/test/java/com/engineering/thesis/backend/ServiceUnitTests/Authentification/Login.java
@@ -1,0 +1,4 @@
+package com.engineering.thesis.backend.ServiceUnitTests.Authentification;
+
+public class Login {
+}
