+++ backend/src/test/java/com/engineering/thesis/backend/ServiceUnitTests/Authentification/Register.java
@@ -1,0 +1,4 @@
+package com.engineering.thesis.backend.ServiceUnitTests.Authentification;
+
+public class Register {
+}
