+++ backend/src/test/java/com/engineering/thesis/backend/ServiceUnitTests/Doctor/GetById.java
@@ -1,0 +1,4 @@
+package com.engineering.thesis.backend.ServiceUnitTests.Doctor;
+
+public class GetById {
+}
