+++ backend/src/test/java/com/engineering/thesis/backend/ServiceUnitTests/MedicalFacility/Post.java
@@ -1,0 +1,4 @@
+package com.engineering.thesis.backend.ServiceUnitTests.MedicalFacility;
+
+public class Post {
+}
