+++ backend/src/test/java/com/engineering/thesis/backend/ServiceUnitTests/PatientMedicalData/GetById.java
@@ -1,0 +1,4 @@
+package com.engineering.thesis.backend.ServiceUnitTests.PatientMedicalData;
+
+public class GetById {
+}
