+++ backend/src/test/java/com/engineering/thesis/backend/ServiceUnitTests/Price/GetById.java
@@ -1,0 +1,5 @@
+package com.engineering.thesis.backend.ServiceUnitTests.Price;
+
+public class GetById {
+
+}
