+++ backend/src/test/java/com/engineering/thesis/backend/ServiceUnitTests/Price/Post.java
@@ -1,0 +1,4 @@
+package com.engineering.thesis.backend.ServiceUnitTests.Price;
+
+public class Post {
+}
