+++ backend/src/test/java/com/engineering/thesis/backend/ServiceUnitTests/User/Get.java
@@ -1,0 +1,4 @@
+package com.engineering.thesis.backend.ServiceUnitTests.User;
+
+public class Get {
+}
